Move pool tests to separate file and yarn target
